src/test/java/hello/servlert/domain/member/MemberRepositoryTest.java
package hello.servlert.domain.member;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class MemberRepositoryTest {
    // JUNIT5 부터는 별도의 접근제한자가 필요 X

    //MemberRepository는 싱글톤이라 new X
    //하지만 스프링을 쓰면 신경 쓸 필요 없음. 알아서 싱글톤을 유지해 줌
    MemberRepository memberRepository = MemberRepository.getInstance();

    @AfterEach
    void afterEach() {
        memberRepository.clearStore();
    }

    //매 테스트 마다 실행되는 메서드 save find 가 연달아 실행되다 보면 테스트가 꼬일수 있다

    @Test
    void save() {

        //given

        Member member = new Member("hello", 20);

        //when
        Member savedMember = memberRepository.save(member);

        //then
        Member findMember = memberRepository.findById(savedMember.getId());
        Assertions.assertThat(findMember).isEqualTo(savedMember);

    }

    @Test
    void findAll() {
        //given
        Member member1 = new Member("member1", 20);
        Member member2 = new Member("member2", 30);

        memberRepository.save(member1);
        memberRepository.save(member2);

        //when
        List<Member> result = memberRepository.findAll();

        //then
        Assertions.assertThat(result.size()).isEqualTo(2);
        Assertions.assertThat(result).contains(member1, member2);




    }





}